Modification for the header section of the page.
Now looks good and as for today I don't see any problem with the page.
There is something that bothers me: when making the page smaller in width,
the links at the top move to the left according to the resizing, but the picture doesn't.
It works as I expect to, but I'm not sure if it's visually appealing.