Mofication of header plus completed container-2
Completed the style of the container-2. It's compliant with the project proposal.
However, I'd like to modify this container-2 along with the header+hero in order to
achieve a 2x2 grid with the pictures when shrinking the web. This can be achieve by
means of playing around with the min width in container 1.