Style finished - Small resizing bug container-4
The overall aesthetics of the project is finish. However, there is a small bug
bothering me, which comes when resizing the call to action / container-4 section.
It doesn't keep centered as the other elements --> left margin is kept constant while
right margin decreases; however, web tools points that both margins are the same size...